refactor(ecma/lexer): common `skip_block_comment`
