+++ crates/swc_ecma_parser/src/lexer/mod.rs
@@ -24,7 +24,6 @@ mod jsx;
 mod state;
 mod table;
 mod token;
-pub mod util;
 
 pub(crate) use token::{NextTokenAndSpan, Token, TokenAndSpan, TokenValue};
 
@@ -113,11 +112,6 @@ impl<'a> swc_ecma_lexer::common::lexer::Lexer<'a, TokenAndSpan> for Lexer<'a> {
     #[inline(always)]
     fn atom<'b>(&self, s: impl Into<std::borrow::Cow<'b, str>>) -> swc_atoms::Atom {
         self.atoms.atom(s)
-    }
-
-    #[inline(always)]
-    fn skip_block_comment(&mut self) {
-        self.skip_block_comment();
     }
 
     #[inline(always)]
