+++ crates/swc_ecma_parser/src/lexer/state.rs
@@ -349,8 +349,8 @@ impl swc_ecma_lexer::common::lexer::state::State for State {
     }
 
     #[inline(always)]
-    fn set_had_line_break(&mut self, had_line_break: bool) {
-        self.had_line_break = had_line_break;
+    fn mark_had_line_break(&mut self) {
+        self.had_line_break = true;
     }
 
     #[inline(always)]
